requirements.txt
python-dotenv>=1.0.1
websockets>=12.0
aiohttp>=3.9.5
numpy>=1.26.4
soundfile>=0.12.1
pydub>=0.25.1
opencv-python>=4.10.0.84
click>=8.1.7
openai>=1.40.0
scipy>=1.11.4
